+++ django_project/siteanalytics/views.py
@@ -1,4 +1,5 @@
 # from django.shortcuts import render
+from django.shortcuts import render
 
 def site_analytics_view(request):
     return render(
